rpk/cli: Add 'rpk topic produce', depr. 'rpk api produce'
